my-first-project/requirements.txt
fastapi
uvicorn
onnxruntime-gpu
numpy
pillow
requests
python-multipart
